Remove misleading production profile
Closes gh-333
