+++ src/main/java/org/springframework/samples/petclinic/system/CacheConfig.java
@@ -6,14 +6,9 @@ import javax.cache.configuration.MutableConfiguration;
 import org.springframework.boot.autoconfigure.cache.JCacheManagerCustomizer;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Profile;
 
-/**
- * Cache could be disabled in unit test.
- */
 @org.springframework.context.annotation.Configuration
 @EnableCaching
-@Profile("production")
 class CacheConfig {
 
     @Bean
