+++ src/main/resources/application.properties
@@ -20,6 +20,3 @@ management.endpoints.web.exposure.include=*
 logging.level.org.springframework=INFO
 # logging.level.org.springframework.web=DEBUG
 # logging.level.org.springframework.context.annotation=TRACE
-
-# Active Spring profiles
-spring.profiles.active=production
