--- src/test/java/org/springframework/samples/petclinic/system/ProductionConfigurationTests.java
@@ -1,23 +1,0 @@
-package org.springframework.samples.petclinic.system;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.samples.petclinic.vet.VetRepository;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class ProductionConfigurationTests {
-
-    @Autowired
-    private VetRepository vets;
-
-    @Test
-    public void testFindAll() throws Exception {
-        vets.findAll();
-        vets.findAll(); // served from cache
-    }
-}
